Ajustes en verificación de datos

+++ src/routes/notes.js
@@ -9,41 +9,38 @@ const { isAuthenticated } = require('../helpers/auth.js');
 
 router.get('/notes', isAuthenticated, async (req, res) => {
     try {
-
-        const publications = await Publication.find({ user: req.user.id }).sort({ date: 'desc' }).lean();
-        const publicationsAll = await Publication.find().sort({ date: 'desc' }).lean();
-
         const userAuthenticated = req.user.id;
         const userName = req.user.name;
 
         // possibles errors
         if (!req.user.id || !userAuthenticated || !userName) {
-            req.flash('error_msg', 'Ops! Ocurrió un error');
-            res.render('components/allnotes', { publications: publications, userAuthenticated, publicationsAll, userName });
+            res.redirect('/users/signin');
+            return;
         }
-        //
+
+        const publications = await Publication.find({ user: req.user.id }).sort({ date: 'desc' }).lean();
+        const publicationsAll = await Publication.find().sort({ date: 'desc' }).lean();
 
         res.render('components/allnotes', { publications: publications, userAuthenticated, publicationsAll, userName });
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
+
 });
 
 router.post('/notes/new-note', isAuthenticated, async (req, res) => {
     try {
-
         const { title, description } = req.body;
 
-        const errors = [];
-        if (!title) {
-            errors.push({ text: 'Por favor escribe un título' })
+        if (!title || title.trim() === "") {
+            req.flash('error_msg', 'Por favor escribe un título')
+            res.redirect('/notes')
+            return
         }
-        if (!description) {
-            errors.push({ text: 'Por favor escribe una descripción' })
-        }
-        if (errors.length > 0) {
-            res.render('components/new-publication', { errors, title, description });
+        if (!description || description.trim() === "") {
+            req.flash('error_msg', 'Por favor escribe una descripción')
+            res.redirect('/notes')
+            return
         }
 
         const newPublication = new Publication({ title, description });
@@ -52,8 +49,7 @@ router.post('/notes/new-note', isAuthenticated, async (req, res) => {
         req.flash('success_msg', 'Publicación realizada');
         res.redirect('/notes');
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
 });
 
@@ -62,13 +58,9 @@ router.get('/notes/edit/:id', isAuthenticated, async (req, res) => {
         const userAuthenticated = req.user.id;
         const publication = await Publication.find({ _id: req.params.id }).lean();
 
-        // possibles errors 
-        // 
-
         res.render('components/edit-publication', { publication: publication[0], userAuthenticated });
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
 });
 
@@ -87,54 +79,81 @@ router.get('/notes/public/:id', isAuthenticated, async (req, res) => {
 
         res.render('components/note-public', { publication, userAuthenticated, comments, username, userEmail });
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
 });
 
 router.post('/notes/edit-note/:id', isAuthenticated, async (req, res) => {
     try {
+        if (!req.params.id) {
+            res.redirect('/notes')
+            return;
+        }
+
 
         const { title, description } = req.body;
-        await Publication.findByIdAndUpdate(req.params.id, { title, description });
+        if (!title || title.trim() === "") {
+            req.flash('error_msg', 'Por favor escribe un título');
+            res.redirect(`/notes/edit/${req.params.id}`);
+            return
+        }
+        if (!description || description.trim() === "") {
+            req.flash('error_msg', 'Por favor escribe una descripción');
+            res.redirect(`/notes/edit/${req.params.id}`);
+            return
+        }
 
+        await Publication.findByIdAndUpdate(req.params.id, { title, description });
         req.flash('success_msg', 'Publicación actualizada con éxito')
         res.redirect('/notes');
     } catch (error) {
-        console.error('Error', error);
+        res.status(500).json({ message: 'Ocurrió un error' });
 
     }
 });
 
 router.post('/notes/:_id/like', isAuthenticated, async (req, res) => {
     try {
-        const _id = req.params._id;
+        if (!req.params._id) {
+            res.redirect('/notes')
+            return;
+        }
 
+
+        const _id = req.params._id;
         const publication = await Publication.findOne({ _id });
 
         if (publication) {
             publication.likes = publication.likes + 1;
             await publication.save();
             res.status(200).json({ likes: publication.likes });
-
         }
         else {
-            res.status(500).json({ error: 'Ocurrió un error' });
-
+            res.status(500).json({ message: 'Ocurrió un error' });
         }
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
-
 });
+
+
+
 
 router.post('/notes/:id/comment', isAuthenticated, async (req, res) => {
     try {
+        if (!req.params.id) {
+            res.redirect('/notes')
+            return;
+        }
 
         const { name, email, comment } = req.body;
         const id = req.params.id;
 
+        if (!name || !email || !comment || name.trim() === "" || email.trim() === "" || comment.trim() === "") {
+            req.flash('error_msg', 'Debes agregar un comentario');
+            res.redirect(`/notes/public/${id}`);
+            return;
+        }
 
 
 
@@ -153,32 +172,35 @@ router.post('/notes/:id/comment', isAuthenticated, async (req, res) => {
             );
             await newComment.save();
 
-            req.flash('success_msg', 'Comentario realizado')
+            req.flash('success_msg', 'Comentario realizado');
             res.redirect(`/notes/public/${publication._id}`);
-
+            return;
         }
         else {
             res.redirect('/');
-
+            return;
         }
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
-
 });
+
 
 
 router.post('/notes/delete/:id', isAuthenticated, async (req, res) => {
     try {
-        req.flash('success_msg', 'Publicación eliminada');
+
+        if (!req.params.id) {
+            res.redirect('/notes')
+            return;
+        }
+
         await Publication.findByIdAndDelete(req.params.id);
         await Comment.deleteMany({ publication_id: req.params.id });
+        req.flash('success_msg', 'Publicación eliminada');
         res.redirect('/notes');
-
     } catch (error) {
-        console.error('Error', error);
-
+        res.status(500).json({ message: 'Ocurrió un error' });
     }
 });
